Move `cvd/unittests/selector` tests to `cvd/cli/selector`
Bug: b/416769281
