+++ base/cvd/cuttlefish/host/commands/cvd/cli/selector/group_record_test.cpp
