+++ base/cvd/cuttlefish/host/commands/cvd/cli/selector/instance_database_helper.cpp
@@ -13,7 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "cuttlefish/host/commands/cvd/unittests/selector/instance_database_helper.h"
+#include "cuttlefish/host/commands/cvd/cli/selector/instance_database_helper.h"
 
 #include <cstdlib>
