+++ base/cvd/cuttlefish/host/commands/cvd/cli/selector/instance_database_test.cpp
@@ -20,8 +20,8 @@
 #include "cuttlefish/common/libs/utils/files.h"
 #include "cuttlefish/common/libs/utils/json.h"
 #include "cuttlefish/common/libs/utils/result_matchers.h"
+#include "cuttlefish/host/commands/cvd/cli/selector/instance_database_helper.h"
 #include "cuttlefish/host/commands/cvd/instances/instance_database.h"
-#include "cuttlefish/host/commands/cvd/unittests/selector/instance_database_helper.h"
 
 /*
  * SetUp creates a mock ANDROID_HOST_OUT directory where there is
