+++ base/cvd/cuttlefish/host/commands/cvd/cli/selector/parser_ids_test.cpp
@@ -17,7 +17,7 @@
 #include <gtest/gtest.h>
 
 #include "cuttlefish/host/commands/cvd/cli/selector/start_selector_parser.h"
-#include "cuttlefish/host/commands/cvd/unittests/selector/parser_ids_helper.h"
+#include "cuttlefish/host/commands/cvd/cli/selector/parser_ids_helper.h"
 
 namespace cuttlefish {
 namespace selector {
