+++ base/cvd/cuttlefish/host/commands/cvd/cli/selector/parser_names_test.cpp
@@ -19,7 +19,7 @@
 #include "cuttlefish/common/libs/utils/result_matchers.h"
 #include "cuttlefish/host/commands/cvd/cli/selector/selector_common_parser.h"
 #include "cuttlefish/host/commands/cvd/cli/selector/start_selector_parser.h"
-#include "cuttlefish/host/commands/cvd/unittests/selector/parser_names_helper.h"
+#include "cuttlefish/host/commands/cvd/cli/selector/parser_names_helper.h"
 
 namespace cuttlefish {
 namespace selector {
